+++ Editor/Aseprite/Chunks/CellExtra.cs
@@ -2,11 +2,14 @@ using System.IO;
 
 namespace UnityEditor.U2D.Aseprite
 {
-    /// <summary>Not supported yet.</summary>
+    /// <summary>
+    /// Parsed representation of an Aseprite CellExtra chunk.
+    /// </summary>
+    /// <note>Not supported yet.</note>
     internal class CellExtra : BaseChunk
     {
         public override ChunkTypes chunkType => ChunkTypes.CellExtra;
-        public CellExtra(uint chunkSize) : base(chunkSize) { }
+        internal CellExtra(uint chunkSize) : base(chunkSize) { }
         protected override void InternalRead(BinaryReader reader) { }
     }
 }