+++ Editor/Aseprite/Chunks/ExternalFilesChunk.cs
@@ -2,12 +2,15 @@ using System.IO;
 
 namespace UnityEditor.U2D.Aseprite
 {
-    /// <summary>Not supported yet.</summary>
+    /// <summary>
+    /// Parsed representation of an Aseprite ExternalFiles chunk.
+    /// </summary>
+    /// <note>Not supported yet.</note>
     internal class ExternalFilesChunk : BaseChunk
     {
         public override ChunkTypes chunkType => ChunkTypes.ExternalFiles;
 
-        public ExternalFilesChunk(uint chunkSize) : base(chunkSize) { }
+        internal ExternalFilesChunk(uint chunkSize) : base(chunkSize) { }
         protected override void InternalRead(BinaryReader reader) { }
     }
 }