+++ Editor/Aseprite/Chunks/MaskChunk.cs
@@ -2,12 +2,15 @@ using System.IO;
 
 namespace UnityEditor.U2D.Aseprite
 {
-    /// <summary>Not supported yet.</summary>
+    /// <summary>
+    /// Parsed representation of an Aseprite Mask chunk.
+    /// </summary>
+    /// <note>Not supported yet.</note>
     internal class MaskChunk : BaseChunk
     {
         public override ChunkTypes chunkType => ChunkTypes.Mask;
 
-        public MaskChunk(uint chunkSize) : base(chunkSize) { }
+        internal MaskChunk(uint chunkSize) : base(chunkSize) { }
         protected override void InternalRead(BinaryReader reader) { }
     }
 }