+++ Editor/Aseprite/Chunks/NoneChunk.cs
@@ -2,10 +2,12 @@ using System.IO;
 
 namespace UnityEditor.U2D.Aseprite
 {
-    /// <summary>Not supported yet.</summary>
+    /// <summary>
+    /// Empty default chunk.
+    /// </summary>
     internal class NoneChunk : BaseChunk
     {
-        public NoneChunk(uint chunkSize) : base(chunkSize) { }
+        internal NoneChunk(uint chunkSize) : base(chunkSize) { }
         protected override void InternalRead(BinaryReader reader) { }
     }
 }