+++ Editor/Aseprite/Chunks/OldPaletteChunk.cs
@@ -2,12 +2,15 @@ using System.IO;
 
 namespace UnityEditor.U2D.Aseprite
 {
-    /// <summary>Not supported yet.</summary>
+    /// <summary>
+    /// Parsed representation of an Aseprite Old Palette chunk.
+    /// </summary>
+    /// <note>Not supported yet.</note>
     internal class OldPaletteChunk : BaseChunk
     {
         public override ChunkTypes chunkType => ChunkTypes.OldPalette;
         
-        public OldPaletteChunk(uint chunkSize) : base(chunkSize) { }
+        internal OldPaletteChunk(uint chunkSize) : base(chunkSize) { }
         protected override void InternalRead(BinaryReader reader) { }
     }
 }