+++ Editor/Aseprite/Chunks/PathChunk.cs
@@ -2,12 +2,15 @@ using System.IO;
 
 namespace UnityEditor.U2D.Aseprite
 {
-    /// <summary>Not supported yet.</summary>
+    /// <summary>
+    /// Parsed representation of an Aseprite Path chunk.
+    /// </summary>
+    /// <note>Not supported yet.</note>
     internal class PathChunk : BaseChunk
     {
         public override ChunkTypes chunkType => ChunkTypes.Path;
 
-        public PathChunk(uint chunkSize) : base(chunkSize) { }
+        internal PathChunk(uint chunkSize) : base(chunkSize) { }
         protected override void InternalRead(BinaryReader reader) { }
     }
 }